Default export for Navigation and Auth
